+++ src/common/editor/code.editor.ts
@@ -2,4 +2,4 @@ export { SetupEditor, MonacoCodeEditor } from './monaco';
 export { VirtualEditor, FileOpenValueEditor, FileBrowseValueEditor, ExternalExitValueEditor, ItemsValueEditor, RoomExit, RoomStates } from './virtual.editor';
 export { AreaDesigner, RoomTypes, MonsterTypes, StdObjectType } from './area.designer';
 export { DebugTimer, Source, FileState } from './editor.base';
-export { createFunction, formatFunctionPointer, formatArgumentList } from './lpc';
+export { createFunction, formatFunctionPointer, formatArgumentList, formatMapping, LPCFormatter } from './lpc';